BioMind/requirements.txt
# Core Google Cloud dependencies
google-cloud-aiplatform==1.71.1  # Includes Vertex AI Matching Engine functionality
google-cloud-storage>=2.13.0
google-cloud-bigquery>=3.10.0
google-cloud-documentai>=2.25.0
vertexai==1.71.1

# Google GenAI SDK
google-generativeai>=0.3.0
google-genai>=0.3.0

# Data processing and analysis
pandas>=2.1.3
numpy>=1.24.0
biopython>=1.81

# API connectors
requests>=2.31.0
beautifulsoup4>=4.12.2
urllib3>=2.0.7
feedparser>=6.0.11
lxml>=5.2.1

# LangChain and embeddings
langchain>=0.0.350
langchain-community>=0.0.10
langchain-google-vertexai>=0.0.10
sentence-transformers>=2.2.2

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0
streamlit>=1.37.0
pypdf>=4.2.0
python-docx>=1.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-multipart>=0.0.7
tqdm>=4.66.4

# Optional: Development and testing
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
